+++ src/drawer.multicanvas.js
@@ -156,6 +156,9 @@ export default class MultiCanvas extends Drawer {
 
             entry.clearWave();
         });
+
+        this.style(this.progressWave, { width: totalWidth + "px" });
+        this.style(this.progressWave, { "clip-path": this.makeInset(totalWidth)});
     }
 
     /**
@@ -556,11 +559,26 @@ export default class MultiCanvas extends Drawer {
     }
 
     /**
+     * build a css inset string for masking off portions of the progessWave
+     *
+     * In order to avoid browser layout passes, we leave our progress wave at full width
+     * but mask a portion of it off using the `clip-path` CSS property.
+     *
+     * @param {number} rightInset=number of pixels to clip off the right
+     * @return {string} css
+     */
+    makeInset(rightInset) {
+        return `inset(0px ${rightInset}px 0px 0px)`;
+    }
+
+    /**
      * Render the new progress
      *
      * @param {number} position X-offset of progress position in pixels
      */
     updateProgress(position) {
-        this.style(this.progressWave, { width: position + 'px' });
+        let actualWidth = this.width / this.params.pixelRatio;
+
+        this.style(this.progressWave, { 'clip-path': this.makeInset(actualWidth - position)});
     }
 }